% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_score.R
\name{alignment_score}
\alias{alignment_score}
\title{Compute the alignment score.}
\usage{
alignment_score(al1, al2, match, mismatch, gap)
}
\arguments{
\item{al1}{Aligned sequence 1.}

\item{al2}{Aligned sequence 2.}

\item{match}{An integer which indicates the score for a match between
the two sequences.}

\item{mismatch}{An integer which indicates the score for a mismatch
between the two sequences.}

\item{gap}{An integer which indicates the score for a gap inserted
in one of the two sequences}
}
\value{
The score of the alignment
}
\description{
Compute the alignment score.
}
\examples{

alignment_score('ACGGTTGC','A-GCGT-C', 1, -1, -2)

}
