% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fogsaa_alignment.R
\name{compare}
\alias{compare}
\title{Comparison of two nodes.}
\usage{
compare(new_vertex, current_vertex)
}
\arguments{
\item{new_vertex}{A node}

\item{current_vertex}{A node}
}
\value{
True if the nodes are equal in terms of P1 and P2 and the new node
has an higher score, false in any other case
}
\description{
It compares two nodes based on their pointers P1 and P2 and checking the Tmax
score
}
