% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fogsaa_alignment.R
\name{compute_scores}
\alias{compute_scores}
\title{Computation of present and future score of each node.}
\usage{
compute_scores(new_vertex, current_vertex, match, mismatch, gap)
}
\arguments{
\item{new_vertex}{The node for which we compute the scores}

\item{current_vertex}{The previous node from which we retrieve the PrS score}

\item{match}{An integer which indicates the score for a match between
the two sequences.}

\item{mismatch}{An integer which indicates the score for a mismatch
between the two sequences.}

\item{gap}{An integer which indicates the score for a gap inserted
in one of the two sequences.}
}
\value{
The node in which all the attributes reguarding the different scores
are computed
}
\description{
It computes both the present and future score of each new node, finally it
sums the two to obtain the Fitness score which is used to identify the best
node
}
