% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fogsaa_alignment.R
\name{max_score}
\alias{max_score}
\title{Max score for a node.}
\usage{
max_score(vertex_11, vertex_10, vertex_01)
}
\arguments{
\item{vertex_11}{A node originated from the current node incrementing
both P1 and P2}

\item{vertex_10}{A node originated from the current node incrementing
only P1}

\item{vertex_01}{A node originated from the current node incrementing
only P2}
}
\value{
The node with the highest Fitness score
}
\description{
It computes the node which has the highest score among the three
}
