% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{node}
\alias{node}
\title{A Reference Class to represent the nodes used in the FOGSAA algorithm.}
\description{
The class \code{node} is used to represent objects in the FOGSAA algorithm,
this is the data type on which the algorithm is implemented.
}
\section{Fields}{

\describe{
\item{\code{P1}}{It is the pointer to the characters in sequence 1.}

\item{\code{P2}}{It is the pointer to the characters in sequence 2.}

\item{\code{PrS}}{Present score of the node.}

\item{\code{Fmin}}{Minimum of the future score.}

\item{\code{Fmax}}{Maximum of the future score.}

\item{\code{Tmin}}{Minimum of the Fitted score computed as the sum of PrS and Fmin.}

\item{\code{Tmax}}{Maximum of the Fitted score computed as the sum of PrS and Fmax.}

\item{\code{v1}}{Value of sequence 1 at which P1 is pointing.}

\item{\code{v2}}{Value of sequence 2 at which P2 is pointing.}

\item{\code{x1}}{Defined as n - P2 where n is the length of sequence 2.}

\item{\code{x2}}{Defined as m - P1 where m is the length of sequence 1.}

\item{\code{code}}{Indicates how the instance was originated, 11 if incrementing
both P1 and P2 from the previous node, 01 if only P2 is incremented and 10 if
only P1 is incremented.}
}}


